Office almost done / Soft deleted entities accessibility refactored

+++ OnlineShop.Office.WebApiEndPoint/Controllers/UserManagementControllers/UserController.cs
@@ -2,21 +2,27 @@
 using Microsoft.AspNetCore.Mvc;
 using OnlineShop.Office.Application.Contracts.UserManagement;
 using OnlineShop.Office.Application.Dtos.UserManagementDtos.UserDtos;
+using PublicTools.Constants;
 using PublicTools.Resources;
 
 namespace OnlineShop.Office.WebApiEndPoint.Controllers.UserManagementControllers;
 
 [ApiController]
 [Route("api/User")]
-public class UserController(IUserService userService) : Controller
+public class UserController(IUserService userService, IAuthorizationService authorizationService) : Controller
 {
     private readonly IUserService _userService = userService;
+    private readonly IAuthorizationService _authorizationService = authorizationService;
 
     [Authorize]
     [HttpGet("Get")]
     public async Task<IActionResult> Get([FromBody] GetOnlineShopUserAppDto model)
     {
         if (model is null) return Json(MessageResource.Error_NullInputModel);
+
+        var authorizationResult = await _authorizationService.AuthorizeAsync(User, model, PolicyConstants.OwnerOnlyPolicy);
+        if (!authorizationResult.Succeeded) return Forbid(MessageResource.Error_UnauthorizedOwner);
+
         var getOperationResponse = await _userService.Get(model);
         return getOperationResponse.IsSuccessful ? Ok(getOperationResponse.ResultModel) : Problem(getOperationResponse.ErrorMessage, statusCode: (int)getOperationResponse.HttpStatusCode);
     }
@@ -25,15 +31,20 @@ public class UserController(IUserService userService) : Controller
     public async Task<IActionResult> Post([FromBody] PostOnlineShopUserAppDto model)
     {
         if (model is null) return Json(MessageResource.Error_NullInputModel);
+
         var postOperationResponse = await _userService.Post(model);
         return postOperationResponse.IsSuccessful ? Ok(postOperationResponse.Message) : Problem(postOperationResponse.ErrorMessage, statusCode: (int)postOperationResponse.HttpStatusCode);
     }
 
     [Authorize]
     [HttpPut("EditProfile")]
-    public async Task<IActionResult> Put([FromBody] PutOnlineShopUserPropertiesAppDto model)
+    public async Task<IActionResult> Put([FromBody] PutOnlineShopUserAppDto model)
     {
         if (model is null) return Json(MessageResource.Error_NullInputModel);
+
+        var authorizationResult = await _authorizationService.AuthorizeAsync(User, model, PolicyConstants.OwnerOnlyPolicy);
+        if (!authorizationResult.Succeeded) return Forbid(MessageResource.Error_UnauthorizedOwner);
+
         var postOperationResponse = await _userService.Put(model);
         return postOperationResponse.IsSuccessful ? Ok(postOperationResponse.Message) : Problem(postOperationResponse.ErrorMessage, statusCode: (int)postOperationResponse.HttpStatusCode);
     }
@@ -43,6 +54,10 @@ public class UserController(IUserService userService) : Controller
     public async Task<IActionResult> ChangePassword([FromBody] ChangePasswordAppDto model)
     {
         if (model is null) return Json(MessageResource.Error_NullInputModel);
+
+        var authorizationResult = await _authorizationService.AuthorizeAsync(User, model, PolicyConstants.OwnerOnlyPolicy);
+        if (!authorizationResult.Succeeded) return Forbid(MessageResource.Error_UnauthorizedOwner);
+
         var postOperationResponse = await _userService.ChangePassword(model);
         return postOperationResponse.IsSuccessful ? Ok(postOperationResponse.Message) : Problem(postOperationResponse.ErrorMessage, statusCode: (int)postOperationResponse.HttpStatusCode);
     }
@@ -52,6 +67,10 @@ public class UserController(IUserService userService) : Controller
     public async Task<IActionResult> Delete([FromBody] DeleteOnlineShopUserAppDto model)
     {
         if (model is null) return Json(MessageResource.Error_NullInputModel);
+
+        var authorizationResult = await _authorizationService.AuthorizeAsync(User, model, PolicyConstants.OwnerOnlyPolicy);
+        if (!authorizationResult.Succeeded) return Forbid(MessageResource.Error_UnauthorizedOwner);
+
         var postOperationResponse = await _userService.Delete(model);
         return postOperationResponse.IsSuccessful ? Ok(postOperationResponse.Message) : Problem(postOperationResponse.ErrorMessage, statusCode: (int)postOperationResponse.HttpStatusCode);
     }
